refactor(enrolment-type): EnrolmentType enum moved to constants.
PT-849 PT-1102.
