+++ src/domain/occurrences/occurrencesTable/ActionsDropdown.tsx
@@ -26,7 +26,7 @@ import getDateArray from '../../../utils/getDateArray';
 import { getDomain } from '../../../utils/getDomain';
 import getLocalisedString from '../../../utils/getLocalizedString';
 import { ROUTES } from '../../app/routes/constants';
-import { EnrolmentType } from '../../occurrence/enrolmentInfoFormPart/EnrolmentInfoFormPart';
+import { EnrolmentType } from '../../occurrence/constants';
 import { getPlaceFields } from '../../place/utils';
 import styles from './actionsDropdown.module.scss';
 import CancelOccurrenceModal from './CancelOccurrenceModal';
